# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MNA Vietnam - API Portal
  version: 2.0.0
  description: OpenAPI documentation for backend APIs and Server endpoints of MNA Vietnam.
servers:
  - url: http://localhost:3000/api
    description: Development Server
  - url: https://mnavietnam.vercel.app/api
    description: Staging / Production Server
paths:
...
  /settings:
    get:
      summary: Get Global Settings (Admin)
      description: Fetches dynamic SMTP and notification configurations.
      responses:
        '200':
          description: Returns configuration.
        '401':
          description: Unauthorized.
    put:
      summary: Update Settings (Admin)
      description: Updates SMTP credentials and default notification target emails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Settings updated.
        '401':
          description: Unauthorized.
...